Routing. Task11.
Feat - adding routig to the App.
Refactoring the structure of the project.
