+++ src/components/PageNotFound/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const PageNotFound = () => <h2>Page Not Found</h2>;
+
+export default PageNotFound;
